requirements.txt
discord.py
python-dotenv
PyNaCl
discord-py-interactions